fix contains_datetime
Before, sometimes when extract_datetime didnt agree with the formatting from the utterance it could throw an error. Errors are now caught and a false boolean is returned instead. Meaning that if extract_datetime doesnt like the formatting for the utterance we assume theres no datetime, and just return false instead of crashing in contains_datetime.